+++ src/main/java/io/hhplus/tdd/CustomPointException.java
@@ -1,0 +1,8 @@
+package io.hhplus.tdd;
+
+public class CustomPointException extends RuntimeException{
+    public CustomPointException(String message) {
+        super(message);
+    }
+
+}
