+++ src/main/java/io/hhplus/tdd/point/usecase/PointServiceImpl.java
@@ -1,5 +1,6 @@
 package io.hhplus.tdd.point.usecase;
 
+import io.hhplus.tdd.CustomPointException;
 import io.hhplus.tdd.point.interfaces.repository.PointHistoryRepository;
 import io.hhplus.tdd.point.entities.TransactionType;
 import io.hhplus.tdd.point.interfaces.repository.UserPointRepository;
@@ -28,7 +29,7 @@ public class PointServiceImpl implements PointService {
             UserPoint userPoint = userPointRepository.findById(userId);
             long updatedPoint = userPoint.point() + amount;
             if(updatedPoint>999999L){
-                throw new IllegalArgumentException("999,999 포인트 보유 한도 초과입니다.");
+                throw new CustomPointException("999,999 포인트 보유 한도 초과입니다.");
             }
             userPoint = userPointRepository.insertOrUpdate(userId, updatedPoint);
             pointHistoryRepository.insert(userId, amount, TransactionType.CHARGE, System.currentTimeMillis());
@@ -53,11 +54,11 @@ public class PointServiceImpl implements PointService {
         try{
             UserPoint userPoint = userPointRepository.findById(userId);
             if(amount>500000L){
-                throw new IllegalArgumentException("500,000 포인트 사용 한도 초과입니다.");
+                throw new CustomPointException("500,000 포인트 사용 한도 초과입니다.");
             }
             long updatedPoint = userPoint.point() - amount;
             if(updatedPoint<0){
-                throw new IllegalArgumentException("포인트가 부족합니다.");
+                throw new CustomPointException("포인트가 부족합니다.");
             }else
                 userPoint = userPointRepository.insertOrUpdate(userId, updatedPoint);
             pointHistoryRepository.insert(userId, amount, TransactionType.USE, System.currentTimeMillis());
@@ -67,7 +68,7 @@ public class PointServiceImpl implements PointService {
         }
     }
 
-    /* 유저 포인트 충전/이용 조회 */
+    /* 유저 포인트 이용 내역 조회 */
     @Override
     public List<PointHistory> chargeUsePoints(long userId) {
         return pointHistoryRepository.selectAllByUserId(userId);
